[FIX] adapt to new nilearn api 0.13.0 (#961)
* adapt to new nilearn api 0.13.0

* update code to work with nilearn 0.13.0

* handle new errors

* fix style

* fix broken tests

be more consistent on loading/processing images in float32 where appropriate, and handle nibable headers

* add compatibility with nilearn 0.13.0

* bump the minimum nilearn version to 0.12.0

* bump the dependencies for the minimum version

* remove 3.9 and add 3.14

* label 3.14 support

* fix style

* update minimum python testing version to 3.10

* catch missing usage of python 3.9

* fix number interpretation

* fix import sort

* filter out nans

* filter out voxels with bad data

+++ nimare/tests/test_meta_mkda.py
@@ -146,26 +146,26 @@ def test_KDA_approximate_null(testdata_cbma):
     corr_results = corr.transform(results)
     assert isinstance(results, nimare.results.MetaResult)
     assert isinstance(results.description_, str)
-    assert results.get_map("p", return_type="array").dtype == np.float64
+    assert results.get_map("p", return_type="array").dtype == np.float32
     assert isinstance(corr_results, nimare.results.MetaResult)
     assert isinstance(corr_results.description_, str)
     assert (
         corr_results.get_map(
             "logp_level-voxel_corr-FWE_method-montecarlo", return_type="array"
         ).dtype
-        == np.float64
+        == np.float32
     )
     assert (
         corr_results.get_map(
             "logp_desc-size_level-cluster_corr-FWE_method-montecarlo", return_type="array"
         ).dtype
-        == np.float64
+        == np.float32
     )
     assert (
         corr_results.get_map(
             "logp_desc-mass_level-cluster_corr-FWE_method-montecarlo", return_type="array"
         ).dtype
-        == np.float64
+        == np.float32
     )
 
 
@@ -176,25 +176,25 @@ def test_KDA_fwe_1core(testdata_cbma):
     corr = FWECorrector(method="montecarlo", n_iters=5, n_cores=1)
     corr_results = corr.transform(results)
     assert isinstance(results, nimare.results.MetaResult)
-    assert results.get_map("p", return_type="array").dtype == np.float64
+    assert results.get_map("p", return_type="array").dtype == np.float32
     assert isinstance(corr_results, nimare.results.MetaResult)
     assert (
         corr_results.get_map(
             "logp_level-voxel_corr-FWE_method-montecarlo", return_type="array"
         ).dtype
-        == np.float64
+        == np.float32
     )
     assert (
         corr_results.get_map(
             "logp_desc-mass_level-cluster_corr-FWE_method-montecarlo", return_type="array"
         ).dtype
-        == np.float64
+        == np.float32
     )
     assert (
         corr_results.get_map(
             "logp_desc-size_level-cluster_corr-FWE_method-montecarlo", return_type="array"
         ).dtype
-        == np.float64
+        == np.float32
     )
